+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to use markdown (highlight too) inside quote's format setting and quotes' listings
 - New setting that allows to validate advanced search
 
+### Removed
+- Unnecessary `reload button` setting
+
 ### Fixed
 - Notice from Quote Maker now return `search` rather than `null` from uncreated content